feat(cli): Enable dynamic completions (try to work out build-time generation)
